Add MultiTypeDelegate for management Types
