+++ library/src/main/kotlin/com/drakeet/multitype/OneToManyBuilder.kt
@@ -22,7 +22,7 @@ import androidx.annotation.CheckResult
  * @author Drakeet Xu
  */
 internal class OneToManyBuilder<T>(
-  private val adapter: MultiTypeAdapter,
+  private val adapter: MultiTypeDelegate,
   private val clazz: Class<T>
 ) : OneToManyFlow<T>, OneToManyEndpoint<T> {
